requirements.txt
opencv-python
torch
numpy
torchvision
requests
onnxruntime[gpu]
insightface
supervision
Pillow
matplotlib
numpy
